services/core-api/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Async & Queue
celery==5.3.6
redis==5.0.1

#Environment
python-dotenv==1.0.0

#Utils
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Video Processing
ffmpeg-python==0.2.0
opencv-python-headless==4.8.1.78
pillow==9.5.0
numpy==1.24.3

# AI/ML
ultralytics==8.0.225
torch==2.1.1
torchvision==0.16.1

# Google Cloud
google-cloud-vision==3.4.5

# OCR
rapidocr-onnxruntime

# Audio Processing
openai-whisper
pydub==0.25.1

# NER ( Named Entity Recognition)
transformers==4.36.2

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.1.0

# Vector Database
qdrant-client==1.7.3
sentence-transformers==2.3.1
sentry-sdk[fastapi]==2.22.0
slowapi==0.1.9
